+++ src/main/java/cn/lvshengyi/onelvshengyi/entity/UserDO.java
@@ -7,17 +7,41 @@ import lombok.Data;
  */
 @Data
 public class UserDO {
+
+    /**
+     * 用户id
+     */
     private Integer id;
 
+    /**
+     * 用户名
+     */
     private String username;
 
+    /**
+     * 昵称
+     */
     private String nickname;
 
+    /**
+     * 密码
+     */
     private String password;
 
+    /**
+     * 邮箱
+     */
     private String eMail;
 
-    private Byte status;
+    /**
+     * 用户状态
+     * @see cn.lvshengyi.onelvshengyi.constant.UserStatusEnum
+     */
+    private Integer status;
 
-    private Byte type;
+    /**
+     * 用户类型
+     * @see cn.lvshengyi.onelvshengyi.constant.UserTypeEnum
+     */
+    private Integer type;
 }
